Extract normaliseXml into its own test helper file
