+++ packages/core/tests/helpers/comparison/e2eComparisonTestsHelpers.ts
@@ -7,12 +7,6 @@ import type { DbRecords } from "../../types/ComparisonFile"
 const dbConfig = createDbConfig()
 const sql = postgres(dbConfig)
 
-const normaliseXml = (xml?: string): string =>
-  xml
-    ?.replace(/ Error="HO200200"/g, "")
-    .replace(/ hasError="false"/g, "")
-    .replace(' standalone="yes"', "") ?? ""
-
 const insertRecords = async (records: DbRecords): Promise<void> => {
   const errorList = records.errorList.map((record) => {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -46,4 +40,4 @@ const disconnectDb = async () => {
   await sql.end({ timeout: 5 })
 }
 
-export { clearDatabase, disconnectDb, insertRecords, normaliseXml, sql }
+export { clearDatabase, disconnectDb, insertRecords, sql }
